requirements.txt
mss==9.0.1
numpy==1.26.2
opencv-python==4.8.1.78
pillow==10.1.0
screeninfo==0.8.1
